+++ src/settings/providers/aoai.ts
@@ -1,5 +1,5 @@
-import { getChatStream, type ChatMessage, type OpenAIChatPayload } from "../../openai/chat";
 import type { BaseConnection, BaseCredential, BaseProvider, ChatStreamProxy } from "./base";
+import { getChatStream, type ChatMessage, type OpenAIChatPayload } from "./openai/chat";
 
 export interface AzureOpenAICredential extends BaseCredential {
   id: string;
