+++ src/settings/providers/openai/chat.ts
